backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

elasticsearch==8.14.0
pdfminer.six==20231228
pydantic==2.8.2
python-multipart==0.0.17

google-cloud-aiplatform==1.71.1
google-auth==2.34.0
vertexai==1.71.1

requests==2.32.3
